+++ packages/analytics/analytics-utilities/src/types/chart-types.ts
@@ -4,6 +4,7 @@ export const reportChartTypes = [
   'timeseries_line',
   'choropleth_map',
   'timeseries_bar',
+  'doughnut',
 ] as const
 
 export type ReportChartTypes = typeof reportChartTypes[number]