+++ packages/analytics/dashboard-renderer/src/components/DashboardTile.vue
@@ -22,7 +22,7 @@
         </div>
       </KTooltip>
       <div
-        v-if="canShowKebabMenu || badgeData"
+        v-if="canShowTitleActions"
         class="tile-actions"
         :data-testid="`tile-actions-${tileId}`"
       >
@@ -38,7 +38,7 @@
           @click="editTile"
         />
         <KDropdown
-          v-if="canShowKebabMenu"
+          v-if="canShowKebabMenu && kebabMenuHasItems"
           class="dropdown"
           :data-testid="`chart-action-menu-${tileId}`"
           :kpop-attributes="{ placement: 'bottom-end' }"
@@ -125,6 +125,7 @@ import { MoreIcon, EditIcon } from '@kong/icons'
 import { msToGranularity, type AiExploreAggregations, type AiExploreQuery, type AnalyticsBridge, type ExploreAggregations, type ExploreQuery, type ExploreResultV4, type QueryableAiExploreDimensions, type QueryableExploreDimensions, type TimeRangeV4 } from '@kong-ui-public/analytics-utilities'
 import { CsvExportModal } from '@kong-ui-public/analytics-chart'
 import { TIMEFRAME_LOOKUP } from '@kong-ui-public/analytics-utilities'
+import DoughnutChartRenderer from './DoughnutChartRenderer.vue'
 
 const PADDING_SIZE = parseInt(KUI_SPACE_70, 10)
 
@@ -162,11 +163,21 @@ watch(() => props.definition, async () => {
 }, { immediate: true, deep: true })
 
 const exploreLink = computed(() => {
+  const filters = [...props.context.filters, ...props.definition.query.filters ?? []]
+  const dimensions = props.definition.query.dimensions as QueryableExploreDimensions[] | QueryableAiExploreDimensions[] ?? []
+  // TODO: remove once portal has been added as option in Explore
+  if (filters.some(filter => ('dimension' in filter && filter.dimension === 'portal') ||
+    ('field' in filter && filter.field === 'portal')) ||
+    dimensions.some(dim => dim === 'portal')
+  ) {
+    return ''
+  }
+
   if (queryBridge && queryBridge.exploreBaseUrl) {
     const exploreQuery: ExploreQuery | AiExploreQuery = {
-      filters: [...props.context.filters, ...props.definition.query.filters ?? []],
+      filters: filters,
       metrics: props.definition.query.metrics as ExploreAggregations[] | AiExploreAggregations[] ?? [],
-      dimensions: props.definition.query.dimensions as QueryableExploreDimensions[] | QueryableAiExploreDimensions[] ?? [],
+      dimensions: dimensions,
       time_range: props.definition.query.time_range as TimeRangeV4 || props.context.timeSpec,
       granularity: props.definition.query.granularity || chartDataGranularity.value,
 
@@ -181,7 +192,11 @@ const exploreLink = computed(() => {
 
 const csvFilename = computed<string>(() => i18n.t('csvExport.defaultFilename'))
 
+const canShowTitleActions = computed((): boolean => (canShowKebabMenu.value && (kebabMenuHasItems.value || props.context.editable)) || !!badgeData.value)
+
 const canShowKebabMenu = computed(() => hasKebabMenuAccess && !['golden_signals', 'top_n', 'gauge'].includes(props.definition.chart.type))
+
+const kebabMenuHasItems = computed((): boolean => !!exploreLink.value || ('allowCsvExport' in props.definition.chart && props.definition.chart.allowCsvExport) || props.context.editable)
 
 const rendererLookup: Record<DashboardTileType, Component | undefined> = {
   'timeseries_line': TimeseriesChartRenderer,
@@ -189,6 +204,7 @@ const rendererLookup: Record<DashboardTileType, Component | undefined> = {
   'horizontal_bar': BarChartRenderer,
   'vertical_bar': BarChartRenderer,
   'gauge': SimpleChartRenderer,
+  'doughnut': DoughnutChartRenderer,
   'golden_signals': GoldenSignalsRenderer,
   'top_n': TopNTableRenderer,
   'slottable': undefined,