+++ packages/analytics/analytics-geo-map/CHANGELOG.md
@@ -3,6 +3,49 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.80](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-geo-map@0.2.79...@kong-ui-public/analytics-geo-map@0.2.80) (2025-02-25)
+
+**Note:** Version bump only for package @kong-ui-public/analytics-geo-map
+
+
+
+
+
+## [0.2.79](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-geo-map@0.2.78...@kong-ui-public/analytics-geo-map@0.2.79) (2025-02-24)
+
+**Note:** Version bump only for package @kong-ui-public/analytics-geo-map
+
+
+
+
+
+## [0.2.78](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-geo-map@0.2.77...@kong-ui-public/analytics-geo-map@0.2.78) (2025-02-24)
+
+**Note:** Version bump only for package @kong-ui-public/analytics-geo-map
+
+
+
+
+
+## [0.2.77](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-geo-map@0.2.76...@kong-ui-public/analytics-geo-map@0.2.77) (2025-02-21)
+
+**Note:** Version bump only for package @kong-ui-public/analytics-geo-map
+
+
+
+
+
+## [0.2.76](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-geo-map@0.2.75...@kong-ui-public/analytics-geo-map@0.2.76) (2025-02-21)
+
+
+### Bug Fixes
+
+* **deps:** update all [@kong](https://github.com/kong) scoped dependencies (patch) ([#1960](https://github.com/Kong/public-ui-components/issues/1960)) ([9f510bb](https://github.com/Kong/public-ui-components/commit/9f510bbf1d8fdd3448214f683b997d775e0f6806))
+
+
+
+
+
 ## [0.2.75](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-geo-map@0.2.74...@kong-ui-public/analytics-geo-map@0.2.75) (2025-02-17)