+++ packages/entities/entities-plugins/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [9.12.3](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-plugins@9.12.2...@kong-ui-public/entities-plugins@9.12.3) (2025-02-24)
+
+
+### Bug Fixes
+
+* **forms:** array elements are deleted ([#1977](https://github.com/Kong/public-ui-components/issues/1977)) ([f0f0e8d](https://github.com/Kong/public-ui-components/commit/f0f0e8deed61a6369794fea2cee69674b60ac51f))
+
+
+
+
+
+## [9.12.2](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-plugins@9.12.1...@kong-ui-public/entities-plugins@9.12.2) (2025-02-21)
+
+
+### Bug Fixes
+
+* **deps:** update all [@kong](https://github.com/kong) scoped dependencies (patch) ([#1960](https://github.com/Kong/public-ui-components/issues/1960)) ([9f510bb](https://github.com/Kong/public-ui-components/commit/9f510bbf1d8fdd3448214f683b997d775e0f6806))
+
+
+
+
+
 ## [9.12.1](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-plugins@9.12.0...@kong-ui-public/entities-plugins@9.12.1) (2025-02-19)
 
 **Note:** Version bump only for package @kong-ui-public/entities-plugins